tools/requirements.txt
python3-discogs-client
redis
